{{cookiecutter.project_slug}}/server/requirements/dev.txt
-r base.txt

black==22.1.0
pylint==2.12.2
pylint-django==2.5.0
django-extensions==3.1.5

# Testing
# pytest==7.0.1
# pytest-sugar==0.9.4
# pytest-django==4.5.2
# pytest-xdist[psutil]==2.5.0
# pytest-icdiff==0.5
# factory-boy==3.2.1
